requirements.txt
flask
pymongo
python-dotenv
openai
pycountry
requests
